split classes to prevent headache
